Refactor client into modular classes
